flink_consumer/requirements.txt
flink
kafka-python
pandas
numpy
scikit-learn
statsmodels
influxdb-client
apache-flink